+++ protected/humhub/docs/CHANGELOG.md
@@ -29,6 +29,10 @@ HumHub Change Log
 - Enh: Added isGuestAccessEnabled method in User component
 - Enh: Added flash variable (executeJavascript)to execute js on the next page load
 - Enh: Added possibility to create own file handlers (edit, create, import, export)
+- Enh: Added data-action-process to handle modal processes
+- Enh: Added upload file event (humhub:file:created)
+- Enh: Added custom file handler positions
+- Enh: Moved UploadAction::getFileResponse method to FileHelper::getFileInfos
 
 1.2.0-beta.1 (February 08, 2017)
 --------------------------------